Implement web context menu test. - navi_010, navi_011
